Revert "added the login ui and did the necessary changes to the main code to …"
